+++ src/web.rs
@@ -3,6 +3,7 @@
 use std::borrow::Cow;
 
 pub use crate::json::Json;
+pub use crate::lazy_data::LazyData;
 use crate::Redirect;
 
 /// Create a relative or absolute redirect.